test: remove enzyme
