+++ src/setupTests.js
@@ -27,3 +27,20 @@ Object.defineProperty(window, "Worker", {
   value: Worker,
   writable: true,
 });
+
+const mock = {
+  clearWatch: jest.fn(),
+  getCurrentPosition: jest.fn(),
+  watchPosition: (onSuccess) => {
+    onSuccess({
+      coords: {
+        accuracy: 55,
+        latitude: 47.9913611,
+        longitude: 7.84868,
+      },
+      timestamp: 1552660077044,
+    });
+  },
+};
+
+global.navigator.geolocation = mock;
